requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3>=1.28.0
opensearch-py>=2.0.0
requests>=2.28.0
requests-aws4auth>=1.1.0
pytest>=7.0.0
